requirements.txt
azure-ai-formrecognizer==3.3.0b1
azure-common==1.1.28
azure-core==1.27.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msrest==0.7.1
numpy==1.24.3
oauthlib==3.2.2
pandas==2.0.2
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6
